requirements.txt
flask==3.0.3
boto3==1.35.36
python-dotenv==1.0.1
gunicorn==23.0.0
